# CalculateDistance -> CalculateHelper로 이름바꿈 + 함수 추가
